+++ R/icc.R
@@ -39,6 +39,7 @@
 #'    \item Goldstein H, Browne W, Rasbash J. 2010. Partitioning Variation in Multilevel Models. Understanding Statistics, 1:4, 223-231 (\doi{10.1207/S15328031US0104_02})
 #'    \item Grace-Martion K. The Intraclass Correlation Coefficient in Mixed Models, \href{http://www.theanalysisfactor.com/the-intraclass-correlation-coefficient-in-mixed-models/}{web}
 #'    \item Hox J. 2002. Multilevel analysis: techniques and applications. Mahwah, NJ: Erlbaum
+#'    \item Johnson PC, O'Hara RB. 2014. Extension of Nakagawa & Schielzeth's R2GLMM to random slopes models. Methods Ecol Evol, 5: 944-946. (\doi{10.1111/2041-210X.12225})
 #'    \item Nakagawa S, Johnson P, Schielzeth H (2017) The coefficient of determination R2 and intra-class correlation coefficient from generalized linear mixed-effects models revisted and expanded. J. R. Soc. Interface 14. \doi{10.1098/rsif.2017.0213}
 #'    \item Rabe-Hesketh S, Skrondal A. 2012. Multilevel and longitudinal modeling using Stata. 3rd ed. College Station, Tex: Stata Press Publication
 #'    \item Raudenbush SW, Bryk AS. 2002. Hierarchical linear models: applications and data analysis methods. 2nd ed. Thousand Oaks: Sage Publications
@@ -99,6 +100,11 @@
 #'    nonetheless, but it is usually no meaningful summary of the
 #'    proportion of variances.
 #'    \cr \cr
+#'    To get a meaningful ICC also for models with random slopes, use \code{adjusted = TRUE}.
+#'    The adjusted ICC used the mean random effect variance, which is based
+#'    on the random effect variances for each value of the random slope
+#'    (see \cite{Johnson 2014}).
+#'    \cr \cr
 #'    \strong{ICC for models with multiple or nested random effects}
 #'    \cr \cr
 #'    \strong{Caution:} By default, for three-level-models, depending on the
@@ -115,7 +121,8 @@
 #'    The latter also takes the fixed effects variances into account (see
 #'    \cite{Nakagawa et al. 2017}). If random effects are not nested and not
 #'    cross-classified, the adjusted (\code{adjusted = TRUE}) and unadjusted
-#'    (\code{adjusted = FALSE}) ICC are identical.
+#'    (\code{adjusted = FALSE}) ICC are identical. \code{adjust = TRUE} returns
+#'    a meaningful ICC for models with random slopes.
 #'    \cr \cr
 #'    \strong{ICC for specific group-levels}
 #'    \cr \cr
@@ -160,9 +167,10 @@
 #' icc(fit0)
 #'
 #' # note: ICC for random-slope-intercept model usually not
-#' # meaningful - see 'Note'.
+#' # meaningful, unless you use "adjusted = TRUE" - see 'Note'.
 #' fit1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
 #' icc(fit1)
+#' icc(fit1, adjusted = TRUE)
 #'
 #' sleepstudy$mygrp <- sample(1:45, size = 180, replace = TRUE)
 #' fit2 <- lmer(Reaction ~ Days + (1 | mygrp) + (1 | Subject), sleepstudy)
@@ -217,8 +225,15 @@ icc <- function(x, ...) {
 #' @export
 icc.merMod <- function(x, adjusted = FALSE, ...) {
 
+  add.args <- lapply(match.call(expand.dots = F)$`...`, function(x) x)
+
+  if (obj_has_name(add.args, "type"))
+    type <- add.args[["type"]]
+  else
+    type <- "icc"
+
   # compute adjusted and conditional ICC
-  if (adjusted) return(r2_mixedmodel(x, type = "icc"))
+  if (adjusted) return(r2_mixedmodel(x, type = type))
 
   # get family
   fitfam <- model_family(x)
@@ -271,7 +286,7 @@ icc.merMod <- function(x, adjusted = FALSE, ...) {
   } else {
     resid_var <- switch(
       fitfam$link.fun,
-      log = logVarDist(x, null_model(x), fitfam, sig),
+      log = logVarDist(x, null_model(x), fitfam, sig, type = "ICC"),
       sqrt = 0.25,
       badlink(fitfam$link.fun, fitfam$family)
     )
@@ -358,8 +373,15 @@ icc.merMod <- function(x, adjusted = FALSE, ...) {
 #' @export
 icc.glmmTMB <- function(x, adjusted = FALSE, ...) {
 
+  add.args <- lapply(match.call(expand.dots = F)$`...`, function(x) x)
+
+  if (obj_has_name(add.args, "type"))
+    type <- add.args[["type"]]
+  else
+    type <- "icc"
+
   # compute adjusted and conditional ICC
-  if (adjusted) return(r2_mixedmodel(x, type = "icc"))
+  if (adjusted) return(r2_mixedmodel(x, type = type))
 
   # get family
   fitfam <- model_family(x)
@@ -412,7 +434,7 @@ icc.glmmTMB <- function(x, adjusted = FALSE, ...) {
   } else {
     resid_var <- switch(
       fitfam$link.fun,
-      log = logVarDist(x, null_model(x), fitfam, sig),
+      log = logVarDist(x, null_model(x), fitfam, sig, type = "ICC"),
       sqrt = 0.25,
       badlink(fitfam$link.fun, fitfam$family)
     )
@@ -599,7 +621,7 @@ icc.stanreg <- function(x, re.form = NULL, typical = "mean", prob = .89, ppd = F
       # get slope-intercept-correlations
       rho.01 <- tau.01 / sqrt(tau.00.sums * tau.11.sums)
 
-      message("Caution! ICC for random-slope-intercept models usually not meaningful. See 'Note' in `?icc`.")
+      message("Caution! ICC for random-slope-intercept models usually not meaningful. Use `adjusted = TRUE` to use the mean random effect variance to calculate the ICC. See 'Note' in `?icc`.")
 
     }