+++ R/S3-methods.R
@@ -310,10 +310,19 @@ print.tidy_stan <- function(x, ...) {
 
   if (cumulative) x <- sjmisc::var_rename(x, response.level = "response")
 
-  x$term <- gsub("b_", "", x$term, fixed = TRUE)
+  x$term <- gsub("^b_", "", x$term)
+  x$term <- gsub(pattern = "^bsp_mo", replacement = "", x = x$term)
+  simplex <- string_starts_with(pattern = "simo_mo", x = x$term)
+  x$term <- gsub(
+    pattern = "^simo_mo(.*)(\\.)(.*)(\\.)",
+    replacement = "\\1 \\[\\3\\]",
+    x = x$term
+  )
+  x$simplex <- ""
+  if (!sjmisc::is_empty(simplex))
+    x$simplex[simplex] <- "simplex"
 
   x <- get_hdi_data(x, digits = as.numeric(attr(x, "digits")))
-
 
   # print zero-inflated models ----
 
@@ -338,28 +347,52 @@ print.tidy_stan <- function(x, ...) {
     if (ran.eff) {
       print_stan_ranef(x, zeroinf = TRUE)
     } else {
+      simplex <- which(x$simplex == "simplex")
+
+      if (!sjmisc::is_empty(simplex)) {
+        x.sp <- dplyr::slice(x, !! simplex)
+        x <- dplyr::slice(x, -!! simplex)
+      } else
+        x.sp <- NULL
+
       cat(crayon::blue("## Conditional Model:\n\n"))
       x <- trim_hdi(x)
       colnames(x)[1] <- ""
 
       x %>%
+        dplyr::select(-.data$simplex) %>%
         as.data.frame() %>%
         print(..., row.names = FALSE)
       cat("\n")
+
+      if (!is.null(x.sp))
+        print_stan_simplex(x.sp)
     }
 
 
     if (ran.eff) {
       print_stan_zeroinf_ranef(x.zi)
     } else {
+      simplex <- which(x$simplex == "simplex")
+
+      if (!sjmisc::is_empty(simplex)) {
+        x.sp <- dplyr::slice(x.zi, !! simplex)
+        x.zi <- dplyr::slice(x.zi, -!! simplex)
+      } else
+        x.sp <- NULL
+
       cat(crayon::blue("## Zero-Inflated Model:\n\n"))
       x.zi <- trim_hdi(x.zi)
       colnames(x.zi)[1] <- ""
 
       x.zi %>%
+        dplyr::select(-.data$simplex) %>%
         as.data.frame() %>%
         print(..., row.names = FALSE)
       cat("\n")
+
+      if (!is.null(x.sp))
+        print_stan_simplex(x.sp)
     }
 
 
@@ -388,7 +421,7 @@ print.tidy_stan <- function(x, ...) {
 
         xr <- x %>%
           dplyr::filter(.data$response == !! resp) %>%
-          dplyr::select(-1) %>%
+          dplyr::select(-1, -.data$simplex) %>%
           dplyr::mutate(term = clean_term_name(.data$term)) %>%
           trim_hdi() %>%
           as.data.frame()
@@ -412,7 +445,7 @@ print.tidy_stan <- function(x, ...) {
       cat(crayon::cyan(sprintf("## Residual Correlations\n\n", resp)))
 
       x.cor <- x.cor %>%
-        dplyr::select(-1) %>%
+        dplyr::select(-1, -.data$simplex) %>%
         sjmisc::var_rename(term = "correlation") %>%
         trim_hdi() %>%
         as.data.frame()
@@ -427,11 +460,43 @@ print.tidy_stan <- function(x, ...) {
     print_stan_ranef(x)
 
   } else {
+
+    simplex <- which(x$simplex == "simplex")
+
+    if (!sjmisc::is_empty(simplex)) {
+      x.sp <- dplyr::slice(x, !! simplex)
+      x <- dplyr::slice(x, -!! simplex)
+    } else
+      x.sp <- NULL
+
     colnames(x)[1] <- ""
     x %>%
+      dplyr::select(-.data$simplex) %>%
       as.data.frame() %>%
       print(..., row.names = FALSE)
+
+    if (!is.null(x.sp)) {
+      cat("\n")
+      print_stan_simplex(x.sp, ...)
+    }
   }
+}
+
+
+print_stan_simplex <- function(x, ...) {
+  cat(crayon::blue("## Simplex Parameters:\n\n"))
+  x <- trim_hdi(x)
+
+  if (colnames(x)[1] != "term")
+    x <- dplyr::select(x, -1)
+
+  colnames(x)[1] <- ""
+
+  x %>%
+    dplyr::select(-.data$simplex) %>%
+    as.data.frame() %>%
+    print(..., row.names = FALSE)
+  cat("\n")
 }
 
 
@@ -439,7 +504,7 @@ print.tidy_stan <- function(x, ...) {
 #' @importFrom crayon blue red
 #' @importFrom dplyr slice select filter mutate
 print_stan_ranef <- function(x, zeroinf = FALSE) {
-  # find fixed effects - is type = "all"
+  # find fixed effects - if type = "all"
   fe <- which(x$random.effect == "")
 
   if (!sjmisc::is_empty(fe)) {
@@ -455,17 +520,29 @@ print_stan_ranef <- function(x, zeroinf = FALSE) {
 
     x.fe <- dplyr::slice(x, !! fe)
     x <- dplyr::slice(x, -!! fe)
+
+    simplex <- which(x.fe$simplex == "simplex")
+
+    if (!sjmisc::is_empty(simplex)) {
+      x.sp <- dplyr::slice(x.fe, !! simplex)
+      x.fe <- dplyr::slice(x.fe, -!! simplex)
+    } else
+      x.sp <- NULL
+
     x.fe$term <- clean_term_name(x.fe$term)
     x.fe <- trim_hdi(x.fe)
 
     colnames(x.fe)[2] <- ""
 
     x.fe %>%
-      dplyr::select(-1) %>%
+      dplyr::select(-1, -.data$simplex) %>%
       as.data.frame() %>%
       print(row.names = FALSE)
 
     cat("\n")
+
+    if (!is.null(x.sp))
+      print_stan_simplex(x.sp)
   }
 
   # iterate all random effects
@@ -484,7 +561,7 @@ print_stan_ranef <- function(x, zeroinf = FALSE) {
 
     xr <- x %>%
       dplyr::filter(.data$random.effect == !! r) %>%
-      dplyr::select(-1) %>%
+      dplyr::select(-1, -.data$simplex) %>%
       dplyr::mutate(term = clean_term_name(.data$term)) %>%
       trim_hdi() %>%
       as.data.frame()
@@ -539,7 +616,7 @@ print_stan_mv_re <- function(x, resp) {
 
     xr <- x.fe %>%
       dplyr::filter(.data$response == !! resp) %>%
-      dplyr::select(-1:-2) %>%
+      dplyr::select(-1:-2, -.data$simplex) %>%
       dplyr::mutate(term = clean_term_name(.data$term)) %>%
       trim_hdi() %>%
       as.data.frame()
@@ -600,7 +677,7 @@ print_stan_zeroinf_ranef <- function(x) {
     colnames(x.fe)[2] <- ""
 
     x.fe %>%
-      dplyr::select(-1) %>%
+      dplyr::select(-1, -.data$simplex) %>%
       as.data.frame() %>%
       print(row.names = FALSE)
 
@@ -626,7 +703,7 @@ print_stan_zeroinf_ranef <- function(x) {
 
       xr <- x %>%
         dplyr::filter(.data$random.effect == !! r) %>%
-        dplyr::select(-1) %>%
+        dplyr::select(-1, -.data$simplex) %>%
         dplyr::mutate(term = clean_term_name(.data$term)) %>%
         trim_hdi() %>%
         as.data.frame()