+++ NEWS.md
@@ -1,8 +1,17 @@
 # sjstats 0.17.2
 
+## General
+
+* Revised help-files for `r2()` and `icc()`, also by adding more references.
+
 ## Changes to functions
 
 * `omega_sq()` and `eta_sq()` give more informative messages when using non-supported objects.
+* `r2()` and `icc()` give more informative warnings and messages.
+
+## Bug fixes
+
+* `model_frame()` now correctly names the offset-columns for terms provided as `offset`-argument (i.e. for models where the offset was not specified inside the formula).
 
 # sjstats 0.17.1